Moved lib args parsing functions in cli/lib
